fixes #13
Fixes a race condition in the API calls to add the Vault Client Token Header. A single call involves removal and addition of the client token header.
Between the removal and addition, some other thread could make the API call, resulting in 401 errors. This is because the HttpClient is shared by all threads.
The fallacy happens due to messing with Headers at the HttpClient level. The fix was to compose a per thread HttpRequestMessage and set its headers.