Fixes TOTP key list issue (#180)
This commit fixes a problem with the `TOTP.ReadAllKeysAsync()` method. Running this method always resulted in an InvalidOperationException. Upon looking at the HTTP request that the official vault CLI client sent I noticed that it passes the GET parameter list=true when listing the keys. Backporting this to the library cleared the exception and fixed the code.

Co-authored-by: Yvan Janssens <friedkiwi@qseco.fr>